// src/model/document.ts
export interface Document {
  id: string
  score: number
  values: number[]
  metadata: {
    "web-scraper-order": string
    "web-scraper-start-url": string
    "one-liner": string
    batch: string
    location: string
    type: string
    categories: string
    "profile-href": string
    "url-href": string
    status: string
    founded: string
    description: string
    launches: string
    "company-linkedin-href": string
    "company-twitter-href": string
    "company-crunchbase-href": string
    "company-github": string
    "number-jobs": string
    "team-size": string
    latest_news: string
    "latest_news-href": string
    youtube_link: string
    email: string
    calendar_link: string
    name: string
    founder_count: string
    "founder-1": string
    "founder-1-linkedin-href": string
    "founder-1-twitter-href": string
    "founder-2": string
    "founder-2-linkedin-href": string
    "founder-2-twitter-href": string
    "founder-3": string
    "founder-3-linkedin-href": string
    "founder-3-twitter-href": string
    "founder-4": string
    "founder-4-linkedin-href": string
    "founder-4-twitter-href": string
    "founder-5": string
    "founder-5-linkedin-href": string
    "founder-5-twitter-href": string
    "founder-6": string
    "founder-6-linkedin-href": string
    "founder-6-twitter-href": string
    "founder-7": string
    "founder-7-linkedin-href": string
    "founder-7-twitter-href": string
    "founder-8": string
    "founder-8-linkedin-href": string
    "founder-8-twitter-href": string
  }
}